ConfigManager additions, PLJRSound adjusted + Added option to load Bukkit World, Location and Float using ConfigManager. + Added isFloat to NumberUtil. + Changed volume and pitch in PLJRSound from integer to float.

+++ src/main/java/me/pljr/pljrapi/builders/SoundBuilder.java
@@ -5,19 +5,19 @@ import org.bukkit.Sound;
 
 public class SoundBuilder {
     private Sound sound;
-    private int volume;
-    private int pitch;
+    private float volume;
+    private float pitch;
 
 
     public Sound getSound() {
         return sound;
     }
 
-    public int getVolume() {
+    public float getVolume() {
         return volume;
     }
 
-    public int getPitch() {
+    public float getPitch() {
         return pitch;
     }
 
@@ -41,7 +41,7 @@ public class SoundBuilder {
      * Creates an SoundBuilder with selected sound and volume.
      *
      * @param sound {@link Sound} that will represent the sound.
-     * @param volume int that will represent the volume.
+     * @param volume float that will represent the volume.
      */
     public SoundBuilder(Sound sound, int volume){
         this(sound, volume, 1);
@@ -51,8 +51,8 @@ public class SoundBuilder {
      * Creates an SoundBuilder with selected sound and volume.
      *
      * @param sound {@link Sound} that will represent the sound.
-     * @param volume int that will represent the volume.
-     * @param pitch int that will represent the pitch.
+     * @param volume float that will represent the volume.
+     * @param pitch float that will represent the pitch.
      */
     public SoundBuilder(Sound sound, int volume, int pitch){
         this.sound = sound;
@@ -85,10 +85,10 @@ public class SoundBuilder {
     /**
      * Changes the current volume.
      *
-     * @param volume int that will represent the new volume.
+     * @param volume float that will represent the new volume.
      * @return {@link SoundBuilder} with changed volume.
      */
-    public SoundBuilder withVolume(int volume){
+    public SoundBuilder withVolume(float volume){
         this.volume = volume;
         return this;
     }
@@ -96,10 +96,10 @@ public class SoundBuilder {
     /**
      * Changes the current pitch.
      *
-     * @param pitch int that will represent the new pitch.
+     * @param pitch float that will represent the new pitch.
      * @return {@link SoundBuilder} with changed pitch.
      */
-    public SoundBuilder withPitch(int pitch){
+    public SoundBuilder withPitch(float pitch){
         this.pitch = pitch;
         return this;
     }
